requirements.txt
blinker==1.9.0
cachetools==5.5.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
firebase-admin==6.6.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
Werkzeug==3.1.3